task #3 refactoring

+++ 02. Memory Management/Homework/03. Make Company/MakeCompany.h
@@ -8,7 +8,7 @@ std::shared_ptr<Company> makeCompany(const std::vector<std::string>& properties)
   const auto id = std::stoi(properties.at(0));
   const auto name = properties.at(1);
   std::vector<std::pair<char, char> > employees;
-  for (uint64_t i = 2; i < properties.size(); ++i) {
+  for (size_t i = 2; i < properties.size(); ++i) {
     employees.emplace_back(properties.at(i)[0], properties.at(i)[1]);
   }
   return std::make_shared<Company>(id, name, employees);
